safari-600.7-branch/LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.3_Line_Terminators/S7.3_A6_T3-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Invalid unicode escape in identifier: '\u2028'
S7.3_A6_T3

PASS Expected parsing failure

TEST COMPLETE
